requirements.txt
fastapi
uvicorn[standard]
streamlit
plotly
pandas
textblob
nltk
gensim
python-dotenv
wordcloud
Flask
cryptography
vaderSentiment
requests
twitchio
undetected-chromedriver
selenium==4.10.0
transformers
torch
transformers[torch]
datasets
scikit-learn
evaluate

